Feature/mymap

+++ src/main/java/com/kkosoonnae/jpa/projection/StoreReviewsViewProjection.java
@@ -20,7 +20,7 @@ public record StoreReviewsViewProjection(
         String mainPet
 
 ) {
-    public StoreReviewsResponseDto toDto(Integer averageScope) {
+    public StoreReviewsResponseDto toDto(Double averageScope) {
         return StoreReviewsResponseDto.builder()
                 .storeNo(storeNo)
                 .storeName(storeName)
